Fixed bugs and adding code coverage

See merge request swen225-2018-p1/t01/adventure-game!209

+++ src/mapeditor/MapEditor.java
@@ -133,8 +133,8 @@ public class MapEditor extends Application {
     int rows = 2;
     int cols = 4;
 
-    double itemWidth = this.itemWidth / rows;
-    double itemHeight = this.itemHeight / cols;
+    double itemWidth = (double) this.itemWidth / rows;
+    double itemHeight = (double) this.itemHeight / cols;
 
     for (int i = 0; i < rows; i++) {
       for (int j = 0; j < cols; j++) {
@@ -533,8 +533,8 @@ public class MapEditor extends Application {
 
 
     int roomNum = 3;
-    double roomWidth = boardSize / roomNum;
-    double roomHeight = boardSize / roomNum;
+    double roomWidth = (double) boardSize / roomNum;
+    double roomHeight = (double) boardSize / roomNum;
 
     int tileNum = 10;
     double tileWidth = roomWidth / tileNum;
